requirements.txt
asgiref==3.3.4
astroid==2.5.2
Babel==2.8.0
cairocffi==1.2.0
CairoSVG==2.5.2
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
colorama==0.4.4
cryptography==3.4.7
cssselect2==0.4.1
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.1.7
django-admin-charts==0.23.0
django-admin-shortcuts==2.0.0
django-bower==5.2.0
django-braces==1.14.0
django-cleanup==5.0.0
django-crispy-forms==1.9.2
django-filter==2.4.0
django-heroku==0.3.1
django-intl-tel-input==0.3.1
django-js-asset==1.2.2
django-jsonfield==1.4.1
django-model-utils==4.1.1
django-mptt==0.11.0
django-nvd3==0.9.7
django-phonenumber-field==5.0.0
django-phonenumbers==1.0.1
django-qr-code==2.1.0
django-renderpdf==3.0.1
django-star-ratings==0.9.1
django-weasyprint==1.1.0
djangorestframework==3.12.3
djangorestframework-simplejwt==4.6.0
djcacheutils==3.0.0
gunicorn==20.1.0
html5lib==1.1
idna==2.8
isort==5.8.0
Jinja2==2.11.3
lazy-object-proxy==1.6.0
MarkupSafe==1.1.1
mccabe==0.6.1
mysql==0.0.2
mysqlclient==2.0.1
oauthlib==3.1.0
phonenumbers==8.12.18
Pillow==7.2.0
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.0.1
pylint==2.7.2
Pyphen==0.10.0
python-dateutil==2.8.1
python-decouple==3.4
python-memcached==1.59
python-nvd3==0.14.2
python-slugify==1.1.4
python3-openid==3.2.0
pytz==2021.1
requests==2.22.0
requests-oauthlib==1.3.0
responses==0.13.1
segno==1.3.3
shortuuid==1.0.1
six==1.15.0
sqlparse==0.4.1
stripe==2.55.0
swapper==1.1.2.post1
tinycss2==1.1.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
Unidecode==1.2.0
urllib3==1.25.11
WeasyPrint==52.4
webencodings==0.5.1
whitenoise==5.2.0
wrapt==1.12.1
